python/requirements.txt
chatterbox-tts
flask
flask-cors
torch
torchaudio
scipy
